DRY principle applied and helper function created
